requirements.txt

fastapi
uvicorn
httpx
google-generativeai
openai
pydantic
